refactor(contacts): group drawer components
