--- features/flow/contacts/src/add/native.ts
@@ -1,1 +1,0 @@
-export { ContactsAddContactDrawer } from "./drawer/components/ContactsAddContactDrawer.native";
