+++ features/flow/contacts/src/components/AddContactDrawer/index.native.ts
@@ -1,0 +1,8 @@
+export { ContactsAddContactDrawer } from "./ContactsAddContactDrawer.native";
+export { useAddContactDrawerViewModel } from "./useAddContactDrawerViewModel";
+export type {
+  AddContactDrawerViewModel,
+  ContactsAddContactDrawerLabels,
+  ContactsAddContactDrawerProps,
+  UseAddContactDrawerViewModelOptions,
+} from "./types";
