+++ features/flow/contacts/src/components/AddContactDrawer/index.ts
@@ -1,0 +1,7 @@
+export { useAddContactDrawerViewModel } from "./useAddContactDrawerViewModel";
+export type {
+  AddContactDrawerViewModel,
+  ContactsAddContactDrawerLabels,
+  ContactsAddContactDrawerProps,
+  UseAddContactDrawerViewModelOptions,
+} from "./types";
