+++ features/flow/contacts/src/components/AddContactDrawer/useAddContactDrawerViewModel.ts
@@ -1,5 +1,5 @@
 import { useCallback, useState } from "react";
-import { CONTACT_NAME_MAX_LENGTH } from "../model/constants";
+import { CONTACT_NAME_MAX_LENGTH } from "../../add/model/constants";
 import { useAddContactViewModel } from "../../hooks/useAddContactViewModel";
 import type { AddContactDrawerViewModel, UseAddContactDrawerViewModelOptions } from "./types";
