+++ features/flow/contacts/src/components/index.native.ts
@@ -1,0 +1,3 @@
+export * from "./AddContactDrawer/index.native";
+export * from "./ContactsButton/index.native";
+export * from "./ContactsLedgerSyncIntroduction/index.native";
