+++ features/flow/contacts/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./AddContactDrawer";
+export * from "./ContactsButton";
+export * from "./ContactsLedgerSyncIntroduction";
