+++ features/flow/contacts/src/index.native.ts
@@ -1,11 +1,7 @@
 export * from "./add";
 export * from "./featureIntroduction";
-export * from "./add/native";
 export * from "./featureFlags";
 export * from "./hooks";
 export * from "./list";
-export { ContactsButton } from "./components/ContactsButton/index.native";
-export type { ContactsButtonProps } from "./components/ContactsButton/index.native";
-export { ContactsLedgerSyncIntroductionContent } from "./components/ContactsLedgerSyncIntroduction/index.native";
-export type { ContactsLedgerSyncIntroductionContentProps } from "./components/ContactsLedgerSyncIntroduction/index.native";
+export * from "./components/index.native";
 export * from "./list/native";
