+++ features/flow/contacts/src/index.ts
@@ -3,6 +3,5 @@ export * from "./featureIntroduction";
 export * from "./featureFlags";
 export * from "./hooks";
 export * from "./list";
-export * from "./components/ContactsButton";
-export type { ContactsLedgerSyncIntroductionContentProps } from "./components/ContactsLedgerSyncIntroduction";
+export * from "./components";
 export * from "./list/web";
